requirements.txt
docker-compose
ipython
mysql-connector
flask
